restore ChannelSendReceiveStressTest.kt
